AK: Import Agnostic Kit from the SerenityOS project
This kit is really awesome, so I want to use it in the codebase to
facilitate writing good quality code.

We can't do much yet with the library because everything that requires
allocation doesn't work yet due to not having a working "Heap"
implementation. We should be able to use whatever that doesn't require
to call the new operator though for now.

A tiny StdLib.cpp file was added to address runtime issues that the
compiler requires to do in order to be quiet.