Core+AK+Build System: Make everything more robust
This is a big commit, and although I tried to split it, it seems that
this change goes in one piece like that.

What I changed in the commit:
1. The Linker script of the Core is now more correct. It ensures that
the multiboot header will be at the first 8KB of the image.
2. boot.S (the assembly code that bootstrappes the Core) provide the
pointer of the given info from the Multiboot bootloader, so we can
continue to use it in the Core code.
3. AK code is again stripped to allow the code to be compiled. This is
relevant mostly for the Formatting code in the Agnostic Kit.
4. IO handling code was imported from the SerenityOS project, to
allow sending messages to the debugcon.
5. Assertions are now usuable, so we can assert anywhere in the code.
6. Kmalloc is now usuable too, so we can use the new and delete
operators.
7. The most hard change because lots of tinkering around was with
the CMakeLists files. We can now compile the Core image without any
problem, together with some AK code inside it! :^)