Core: Add initial Memory management singleton
We simply enumerate the provided memory map, and while doing so we
append PhysicalRegion(s) to a Vector in the singleton.

Later on, we will use these PhysicalRegion(s) to generate a memory map,
so it will be suitable for multiple boot protocols for the kernels we
want to support.